trying to add cards

+++ Blackjack.Web/WebSockets/PlayerManager.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Net.WebSockets;
 using Blackjack.Game;
 using Blackjack.Web.Models;
+using Microsoft.Ajax.Utilities;
 using PlayingCards.Domain;
 
 namespace Blackjack.Web.WebSockets
@@ -23,10 +24,10 @@ namespace Blackjack.Web.WebSockets
             CurrentPlayerDto currentPlayerDto = new CurrentPlayerDto
             {
                 GameStatus = game.GameStatus,
-                Dealer = game.Dealer,
+                Dealer = HideDealerSecondCard(game.Dealer),
                 Player = game.Players.FirstOrDefault(p => p.Name == PlayerName),
                 TablePlayers = game.Players.Where(p => p.Name != PlayerName)
-                    .Select(HideTablePlayerCards)
+                    .Select(p => HideTablePlayerCards(p, new []{ 0, 1 }))
                     .ToList(),
                 NewPlayers = game.NewPlayers
             };
@@ -34,26 +35,57 @@ namespace Blackjack.Web.WebSockets
             return currentPlayerDto;
         }
 
-        private Player HideTablePlayerCards(Player player)
+        private Player HideTablePlayerCards(Player player, int[] cardIndexesToHide)
+        {
+            if (player.Game.GameStatus == HandStatus.Done)
+                return player;
+
+            Player playerCopy = CopyPlayer(player);
+
+            List<Card> cardsToHide = cardIndexesToHide.Select(i => new Card
+                {
+                    Number = CardNumber.Blank, Suit = Suit.Blank
+                })
+                .ToList();
+
+            playerCopy.Hand.RemoveRange(0, 2);
+            playerCopy.Hand.InsertRange(0, cardsToHide);
+
+            return playerCopy;
+        }
+
+        private Player HideDealerSecondCard(Player dealer)
+        {
+            if (dealer.Game.GameStatus == HandStatus.Done)
+                return dealer;
+
+            Player dealerCopy = CopyPlayer(dealer);
+
+            Card cardToHide = new Card
+            {
+                Number = CardNumber.Blank, 
+                Suit = Suit.Blank
+            };
+
+            dealerCopy.Hand.RemoveAt(1);
+            dealerCopy.Hand.Insert(1, cardToHide);
+
+            return dealerCopy;
+        }
+
+        private static Player CopyPlayer(Player player)
         {
             Player playerCopy = new Player
             {
                 Name = player.Name,
                 Game = player.Game,
-                Hand = player.Hand.Select(c => new Card { Number = c.Number, Suit = c.Suit}).ToList(),
+                Hand = player.Hand.Select(c => new Card { Number = c.Number, Suit = c.Suit }).ToList(),
                 HandStatus = player.HandStatus,
                 IsTurnToHit = player.IsTurnToHit,
                 Position = player.Position,
                 Score = player.Score,
                 WinningStatus = player.WinningStatus
             };
-
-            List<Card> cardsToHide = playerCopy.Hand.GetRange(0, 2);
-            cardsToHide.ForEach(c => {c.Number = CardNumber.Blank; c.Suit = Suit.Blank;});
-
-            playerCopy.Hand.RemoveRange(0, 2);
-            playerCopy.Hand.InsertRange(0, cardsToHide);
-
             return playerCopy;
         }
     }
